Reorganise headers to make self-contained
