+++ inst/include/dust/cuda/filter_state.hpp
@@ -2,6 +2,7 @@
 #define DUST_CUDA_FILTER_STATE_HPP
 
 #include "dust/filter_state.hpp"
+#include "dust/cuda/device_state.hpp"
 
 namespace dust {
 namespace filter {
