+++ inst/include/dust/interface/dust.hpp
@@ -3,6 +3,7 @@
 
 #include <cstring>
 #include <map>
+
 #include <cpp11/doubles.hpp>
 #include <cpp11/external_pointer.hpp>
 #include <cpp11/integers.hpp>
@@ -11,29 +12,14 @@
 #include <cpp11/raws.hpp>
 #include <cpp11/strings.hpp>
 
-// TODO; this would be neater elsewhere! (fix in #306)
-namespace dust {
-
-template <typename T>
-typename dust::pars_type<T> dust_pars(cpp11::list pars);
-
-template <typename T>
-typename T::data_type dust_data(cpp11::list data);
-
-template <typename T>
-cpp11::sexp dust_info(const dust::pars_type<T>& pars) {
-  return R_NilValue;
-}
-
-}
-
-#include "dust/interface/random.hpp"
-#include "dust/interface/helpers.hpp"
-
 #include "dust/cuda/dust_device.hpp"
-#include "dust/interface/cuda.hpp"
-#include "dust/filter.hpp"
 #include "dust/cuda/filter.hpp"
+#include "dust/dust.hpp"
+#include "dust/filter.hpp"
+
+#include "dust/interface/cuda.hpp"
+#include "dust/interface/helpers.hpp"
+#include "dust/interface/random.hpp"
 
 namespace dust {
